updatehub: Fix installation and probe now settings not being cleared
