+++ updatehub/src/states/entry_point.rs
@@ -4,9 +4,9 @@
 
 use super::{
     actor::{self, SharedState},
-    Park, Poll, Result, State, StateChangeImpl, StateMachine,
+    Park, Poll, Probe, Result, State, StateChangeImpl, StateMachine,
 };
-use slog_scope::debug;
+use slog_scope::{debug, info};
 
 #[derive(Debug, PartialEq)]
 pub(super) struct EntryPoint {}
@@ -38,21 +38,28 @@ impl StateChangeImpl for State<EntryPoint> {
         self,
         shared_state: &mut SharedState,
     ) -> Result<(StateMachine, actor::StepTransition)> {
+        if shared_state.runtime_settings.is_polling_forced() {
+            info!("triggering Probe to finish update.");
+            shared_state.runtime_settings.disable_force_poll()?;
+            return Ok((StateMachine::Probe(self.into()), actor::StepTransition::Immediate));
+        }
+
         // Cleanup temporary settings from last installation
         shared_state.runtime_settings.reset_transient_settings();
 
         if !shared_state.settings.polling.enabled {
-            debug!("Polling is disabled, parking the state machine.");
+            debug!("polling is disabled, parking the state machine.");
             return Ok((StateMachine::Park(self.into()), actor::StepTransition::Immediate));
         }
 
-        debug!("Polling is enabled, moving to Poll state.");
+        debug!("polling is enabled, moving to Poll state.");
         Ok((StateMachine::Poll(self.into()), actor::StepTransition::Immediate))
     }
 }
 
 create_state_step!(EntryPoint => Park);
 create_state_step!(EntryPoint => Poll);
+create_state_step!(EntryPoint => Probe);
 
 #[cfg(test)]
 mod tests {
@@ -98,5 +105,25 @@ mod tests {
             .0;
 
         assert_state!(machine, Poll);
+    }
+
+    #[actix_rt::test]
+    async fn forced_probe() {
+        let settings = Settings::default();
+        let runtime_settings = RuntimeSettings::default();
+        let firmware = Metadata::from_path(&create_fake_metadata(FakeDevice::NoUpdate)).unwrap();
+        let mut shared_state = SharedState { settings, runtime_settings, firmware };
+        shared_state.runtime_settings.reset_installation_settings().unwrap();
+
+        let (machine, trans) = StateMachine::EntryPoint(State(EntryPoint {}))
+            .move_to_next_state(&mut shared_state)
+            .await
+            .unwrap();
+
+        assert_state!(machine, Probe);
+        match trans {
+            actor::StepTransition::Immediate => {}
+            _ => panic!("Unexpected StepTransition: {:?}", trans),
+        }
     }
 }
